Check for reticulate, also remove cache

+++ tests/testthat/helper-redact.R
@@ -20,14 +20,17 @@ redact_port <- function(snapshot) {
 
 
 clean_python_tmp_dir <- function(variables) {
-    python_temp_dir <- dirname(reticulate::py_run_string(
-        "import tempfile; x=tempfile.NamedTemporaryFile().name",
-        local = TRUE
-    )$x)
-    unlink(list.files(
-        python_temp_dir, pattern = "__autograph_generated_file",
-        full.names = TRUE
-    ))
+    if (rlang::is_installed("reticulate")) {
+        python_temp_dir <- dirname(reticulate::py_run_string(
+            "import tempfile; x=tempfile.NamedTemporaryFile().name",
+            local = TRUE
+        )$x)
+        unlink(list.files(
+            python_temp_dir,
+            pattern = "__autograph_generated_file|__pycache__",
+            full.names = TRUE
+        ))
+    }
 }
 
 withr::defer(clean_python_tmp_dir(), teardown_env())
